+++ bindings/java/src/main/rust/lib.rs
@@ -74,6 +74,7 @@ fn build_inliner(
     let inline_style_tags = env.get_bool_field(&cfg, "inlineStyleTags")?;
     let keep_style_tags = env.get_bool_field(&cfg, "keepStyleTags")?;
     let keep_link_tags = env.get_bool_field(&cfg, "keepLinkTags")?;
+    let keep_at_rules = env.get_bool_field(&cfg, "keepAtRules")?;
     let load_remote_stylesheets = env.get_bool_field(&cfg, "loadRemoteStylesheets")?;
     let cache_size = env.get_int_field(&cfg, "cacheSize")?;
     let preallocate_node_capacity = env.get_int_field(&cfg, "preallocateNodeCapacity")?;
@@ -84,6 +85,7 @@ fn build_inliner(
         .inline_style_tags(inline_style_tags)
         .keep_style_tags(keep_style_tags)
         .keep_link_tags(keep_link_tags)
+        .keep_at_rules(keep_at_rules)
         .load_remote_stylesheets(load_remote_stylesheets)
         .extra_css(extra_css.map(Cow::Owned))
         .preallocate_node_capacity(preallocate_node_capacity as usize);