+++ css-inline/src/html/document.rs
@@ -288,9 +288,18 @@ impl Document {
         styles: DocumentStyleMap<'_>,
         keep_style_tags: bool,
         keep_link_tags: bool,
+        at_rules: Option<&String>,
         mode: InliningMode,
     ) -> Result<(), InlineError> {
-        serialize_to(self, writer, styles, keep_style_tags, keep_link_tags, mode)
+        serialize_to(
+            self,
+            writer,
+            styles,
+            keep_style_tags,
+            keep_link_tags,
+            at_rules,
+            mode,
+        )
     }
 
     /// Filter this node iterator to elements matching the given selectors.
@@ -341,6 +350,7 @@ mod tests {
                 IndexMap::default(),
                 false,
                 false,
+                None,
                 InliningMode::Document,
             )
             .expect("Failed to serialize");