+++ css-inline/src/parser.rs
@@ -43,14 +43,7 @@ impl<'i> cssparser::QualifiedRuleParser<'i> for CSSRuleListParser<'_, 'i> {
         _: &ParserState,
         input: &mut cssparser::Parser<'i, 't>,
     ) -> Result<Self::QualifiedRule, cssparser::ParseError<'i, Self::Error>> {
-        // Parse list of declarations
-        let mut parser = CSSDeclarationListParser;
-        let parser = cssparser::RuleBodyParser::new(input, &mut parser);
-        let start = self.0.len();
-        for item in parser.flatten() {
-            self.0.push(item);
-        }
-        Ok((prelude, (start, self.0.len())))
+        Ok((prelude, parse_declarations_into(input, self.0)))
     }
 }
 
@@ -97,4 +90,88 @@ impl<'i> cssparser::QualifiedRuleParser<'i> for CSSDeclarationListParser {
     type Prelude = String;
     type QualifiedRule = Declaration<'i>;
     type Error = ();
+}
+
+pub(crate) struct AtRuleFilteringParser<'d, 'i, 'o> {
+    declarations: &'d mut Vec<Declaration<'i>>,
+    at_rules: &'o mut String,
+}
+
+impl<'d, 'i, 'o> AtRuleFilteringParser<'d, 'i, 'o> {
+    #[inline]
+    pub(crate) fn new(
+        declarations: &'d mut Vec<Declaration<'i>>,
+        at_rules: &'o mut String,
+    ) -> AtRuleFilteringParser<'d, 'i, 'o> {
+        AtRuleFilteringParser {
+            declarations,
+            at_rules,
+        }
+    }
+}
+
+impl<'i> cssparser::QualifiedRuleParser<'i> for AtRuleFilteringParser<'_, 'i, '_> {
+    type Prelude = &'i str;
+    type QualifiedRule = QualifiedRule<'i>;
+    type Error = ();
+
+    fn parse_prelude<'t>(
+        &mut self,
+        input: &mut cssparser::Parser<'i, 't>,
+    ) -> Result<Self::Prelude, cssparser::ParseError<'i, Self::Error>> {
+        Ok(exhaust(input))
+    }
+
+    fn parse_block<'t>(
+        &mut self,
+        prelude: Self::Prelude,
+        _: &ParserState,
+        input: &mut cssparser::Parser<'i, 't>,
+    ) -> Result<Self::QualifiedRule, cssparser::ParseError<'i, Self::Error>> {
+        Ok((prelude, parse_declarations_into(input, self.declarations)))
+    }
+}
+
+impl<'i> cssparser::AtRuleParser<'i> for AtRuleFilteringParser<'_, 'i, '_> {
+    type Prelude = &'i str;
+    type AtRule = QualifiedRule<'i>;
+    type Error = ();
+
+    fn parse_prelude<'t>(
+        &mut self,
+        name: cssparser::CowRcStr<'i>,
+        input: &mut cssparser::Parser<'i, 't>,
+    ) -> Result<Self::Prelude, cssparser::ParseError<'i, Self::Error>> {
+        self.at_rules.push('@');
+        self.at_rules.push_str(&name);
+        Ok(exhaust(input))
+    }
+
+    fn parse_block<'t>(
+        &mut self,
+        prelude: Self::Prelude,
+        _start: &ParserState,
+        input: &mut cssparser::Parser<'i, 't>,
+    ) -> Result<Self::AtRule, cssparser::ParseError<'i, Self::Error>> {
+        let start = self.at_rules.len();
+        self.at_rules.push_str(prelude);
+        self.at_rules.push('{');
+        self.at_rules.push_str(exhaust(input));
+        self.at_rules.push('}');
+        self.at_rules.push(' ');
+        Ok((prelude, (start, self.at_rules.len())))
+    }
+}
+
+fn parse_declarations_into<'i>(
+    input: &mut cssparser::Parser<'i, '_>,
+    declarations: &mut Vec<Declaration<'i>>,
+) -> (usize, usize) {
+    let mut parser = CSSDeclarationListParser;
+    let parser = cssparser::RuleBodyParser::new(input, &mut parser);
+    let start = declarations.len();
+    for item in parser.flatten() {
+        declarations.push(item);
+    }
+    (start, declarations.len())
 }