+++ css-inline/tests/test_inlining.rs
@@ -514,7 +514,10 @@ a {
 
 #[test]
 fn do_not_process_style_tag() {
-    let html = html!("h1 {background-color: blue;}", "<h1>Hello world!</h1>");
+    let html = html!(
+        "@media (max-width: 767px) { padding: 0;} h1 {background-color: blue;}",
+        "<h1>Hello world!</h1>"
+    );
     let options = InlineOptions {
         inline_style_tags: false,
         keep_style_tags: true,
@@ -524,13 +527,16 @@ fn do_not_process_style_tag() {
     let result = inliner.inline(&html).unwrap();
     assert_eq!(
         result,
-        "<html><head><style>h1 {background-color: blue;}</style></head><body><h1>Hello world!</h1></body></html>"
+        "<html><head><style>@media (max-width: 767px) { padding: 0;} h1 {background-color: blue;}</style></head><body><h1>Hello world!</h1></body></html>"
     )
 }
 
 #[test]
-fn do_not_process_style_tag_and_remove() {
-    let html = html!("h1 {background-color: blue;}", "<h1>Hello world!</h1>");
+fn do_not_process_and_remove_style_tag() {
+    let html = html!(
+        "@media (max-width: 767px) { padding: 0;} h1 {background-color: blue;}",
+        "<h1>Hello world!</h1>"
+    );
     let options = InlineOptions {
         keep_style_tags: false,
         inline_style_tags: false,
@@ -541,6 +547,26 @@ fn do_not_process_style_tag_and_remove() {
     assert_eq!(
         result,
         "<html><head></head><body><h1>Hello world!</h1></body></html>"
+    )
+}
+
+#[test]
+fn do_not_process_and_remove_style_tag_but_keep_at_rules() {
+    let html = html!(
+        "@media (max-width: 767px) { padding: 0;} h1 {background-color: blue;}",
+        "<h1>Hello world!</h1>"
+    );
+    let options = InlineOptions {
+        keep_style_tags: false,
+        inline_style_tags: false,
+        keep_at_rules: true,
+        ..Default::default()
+    };
+    let inliner = CSSInliner::new(options);
+    let result = inliner.inline(&html).unwrap();
+    assert_eq!(
+        result,
+        "<html><head><style>@media (max-width: 767px) { padding: 0;} </style></head><body><h1>Hello world!</h1></body></html>"
     )
 }
 
@@ -584,7 +610,10 @@ fn invalid_rule(style: &str, expected: &str) {
 
 #[test]
 fn remove_style_tag() {
-    let html = html!("h1 {background-color: blue;}", "<h1>Hello world!</h1>");
+    let html = html!(
+        "@media (max-width: 600px) { h1 { font-size: 18px; } }\nh1 {background-color: blue;}",
+        "<h1>Hello world!</h1>"
+    );
     let result = inline(&html).unwrap();
     assert_eq!(result, "<html><head></head><body><h1 style=\"background-color: blue;\">Hello world!</h1></body></html>")
 }
@@ -598,6 +627,7 @@ fn remove_multiple_style_tags() {
 h1 {
     text-decoration: none;
 }
+@media (max-width: 600px) { h1 { font-size: 18px; } }
 </style>
 <style>
 .test-class {
@@ -624,6 +654,45 @@ a {
 <body>
 <a class="test-class" href="https://example.com" style="color: #ffffff;">Test</a>
 <h1 style="text-decoration: none;">Test</h1>
+
+
+    </body></html>"#
+    )
+}
+
+#[test]
+fn keep_multiple_at_rules() {
+    let html = r#"
+<html>
+<head>
+<style>
+@media (max-width: 600px) { h1 { font-size: 18px; } }
+@media (max-width: 400px) { h1 { font-size: 12px; } }
+</style>
+<style>
+@media (max-width: 200px) { h1 { font-size: 8px; } }
+</style>
+</head>
+<body>
+<h1>Test</h1>
+</body>
+</html>
+    "#;
+
+    let options = InlineOptions {
+        keep_at_rules: true,
+        ..Default::default()
+    };
+    let inliner = CSSInliner::new(options);
+    let result = inliner.inline(html).unwrap();
+    assert_eq!(
+        result,
+        r#"<html><head><style>@media (max-width: 600px) { h1 { font-size: 18px; } } @media (max-width: 400px) { h1 { font-size: 12px; } } @media (max-width: 200px) { h1 { font-size: 8px; } } </style>
+
+
+</head>
+<body>
+<h1>Test</h1>
 
 
     </body></html>"#
@@ -907,6 +976,7 @@ fn keep_style_tags() {
 <html>
 <head>
 <style>
+@media (max-width: 600px) { h1 { font-size: 18px; } }
 h2 { color: red; }
 </style>
 </head>
@@ -915,7 +985,7 @@ h2 { color: red; }
 </body>
 </html>"#;
     let inlined = inliner.inline(html).unwrap();
-    assert_eq!(inlined, "<html><head>\n<style>\nh2 { color: red; }\n</style>\n</head>\n<body>\n<h2 style=\"color: red;\"></h2>\n\n</body></html>");
+    assert_eq!(inlined, "<html><head>\n<style>\n@media (max-width: 600px) { h1 { font-size: 18px; } }\nh2 { color: red; }\n</style>\n</head>\n<body>\n<h2 style=\"color: red;\"></h2>\n\n</body></html>");
 }
 
 #[test]
@@ -938,6 +1008,27 @@ fn keep_link_tags() {
         inlined,
         "<html><head>\n<link href=\"external.css\" rel=\"stylesheet\">\n</head>\n<body>\n<h1 style=\"color: blue;\"></h1>\n\n</body></html>",
     );
+}
+
+#[test]
+fn keep_at_rules() {
+    let inliner = CSSInliner::options().keep_at_rules(true).build();
+    let html = r#"
+<html>
+<head>
+<style>
+h1 { color: blue; }
+@media (max-width: 600px) { h1 { font-size: 18px; } }
+p { margin: 10px; }
+</style>
+</head>
+<body>
+<h1>Hello</h1><p>World</p>
+</body>
+</html>"#;
+    let inlined = inliner.inline(html).unwrap();
+    let expected = "<html><head><style>@media (max-width: 600px) { h1 { font-size: 18px; } } </style>\n\n</head>\n<body>\n<h1 style=\"color: blue;\">Hello</h1><p style=\"margin: 10px;\">World</p>\n\n</body></html>";
+    assert_eq!(inlined, expected);
 }
 
 #[test]
@@ -1070,7 +1161,7 @@ fn test_disable_cache() {
         ))
         .cache(None)
         .build();
-    let debug = format!("{:?}", inliner);
+    let debug = format!("{inliner:?}");
     assert_eq!(debug, "CSSInliner { options: InlineOptions { inline_style_tags: true, keep_style_tags: false, keep_link_tags: false, base_url: None, load_remote_stylesheets: true, cache: None, extra_css: None, preallocate_node_capacity: 32, .. } }");
 }